requirements-test.txt
pytest
pytest-asyncio
pytest-cov
pytest-mock
httpx
factory-boy
faker
pytest-env
freezegun
